Theme lists UI, tempfix boost button
Since I don't use lists, I've overlooked theming a few elements like the add-a-list namer and the background for an empty list. Also, I've fixed the inactive coloration on the boost button, but in a way that breaks the boost animation. I need to learn more about SVG parameters before I can fix it in a way that preserves the animation.