support profile fields
make them half-transparent black for now... could use bg or bgPage in the future, but there's some extra weirdness going on with the account header and the way it's implemented in vanilla mastodon. it doesn't really look good -- it's like 90% gray overlay. i'll come back to this if/when upstream mastodon redoes their account headers.