Bring compose-form in front of inner drawer
long replies with lots of text will cause the compose form to stretch into the space dedicated for elephant friend, and elephant friend will get in front of the compose box, making it hard to see what you're typing. this commit properly defines that the compose form should be above the inner drawer, and therefore above elephant friend.